add logic for face, fix bug
