+++ click/predict.py
@@ -1,5 +1,5 @@
 import click
-from s3vae.train import run
+from s3vae.train import run_prototype
 from s3vae.utils import load_config
 from pkg_resources import resource_filename
 
@@ -8,11 +8,11 @@ def cli():
     pass
 
 @cli.command(short_help='Generate moving mnist with s3vae')
-@click.option('--config_filename', '-c', type=click.STRING, help='config filename')
-def generate_moving_mnist(config_filename):
+# @click.option('--config_filename', '-c', type=click.STRING, help='config filename')
+def generate_sequential_image():
+    config_filename = resource_filename(__name__, '../s3vae_config.yaml')
     config = load_config(config_filename)
-    data_dir = config['filename']
-    run(config, data_dir)
+    run_prototype(config)
 
 
 if __name__ == '__main__':
